+++ src/Items/NullItem.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Swis\JsonApi\Client\Items;
-
-class NullItem extends JenssegersItem
-{
-}