+++ src/Relations/HasOneRelation.php
@@ -3,13 +3,13 @@
 namespace Swis\JsonApi\Client\Relations;
 
 use Swis\JsonApi\Client\Interfaces\DataInterface;
+use Swis\JsonApi\Client\Interfaces\ItemInterface;
 use Swis\JsonApi\Client\Interfaces\RelationInterface;
-use Swis\JsonApi\Client\Items\JenssegersItem;
 
 class HasOneRelation implements RelationInterface
 {
     /**
-     * @var \Swis\JsonApi\Client\Items\JenssegersItem
+     * @var \Swis\JsonApi\Client\Interfaces\ItemInterface
      */
     protected $included;
 
@@ -24,7 +24,7 @@ class HasOneRelation implements RelationInterface
     protected $id;
 
     /**
-     * @var \Swis\JsonApi\Client\Items\JenssegersItem
+     * @var \Swis\JsonApi\Client\Interfaces\ItemInterface
      */
     protected $parentItem;
 
@@ -34,10 +34,10 @@ class HasOneRelation implements RelationInterface
     protected $omitIncluded = false;
 
     /**
-     * @param string                                    $type
-     * @param \Swis\JsonApi\Client\Items\JenssegersItem $item
+     * @param string                                        $type
+     * @param \Swis\JsonApi\Client\Interfaces\ItemInterface $item
      */
-    public function __construct(string $type, JenssegersItem $item)
+    public function __construct(string $type, ItemInterface $item)
     {
         $this->parentItem = $item;
         $this->type = $type;
@@ -52,8 +52,10 @@ class HasOneRelation implements RelationInterface
      */
     public function associate(DataInterface $included)
     {
-        if (!$included instanceof JenssegersItem) {
-            throw new \InvalidArgumentException('HasOne expects relation to be a JenssegersItem');
+        if (!$included instanceof ItemInterface) {
+            throw new \InvalidArgumentException(
+                sprintf('HasOne expects relation to be an instance of %s', ItemInterface::class)
+            );
         }
 
         $this->setId($included->getId());
@@ -119,7 +121,7 @@ class HasOneRelation implements RelationInterface
     }
 
     /**
-     * @return \Swis\JsonApi\Client\Items\JenssegersItem|null
+     * @return \Swis\JsonApi\Client\Interfaces\ItemInterface|null
      */
     public function getIncluded()
     {