+++ tests/_mocks/Items/AnotherRelatedItem.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Swis\JsonApi\Client\Tests\Mocks\Items\Jenssegers;
+namespace Swis\JsonApi\Client\Tests\Mocks\Items;
 
-use Swis\JsonApi\Client\Items\JenssegersItem;
+use Swis\JsonApi\Client\Item;
 
-class AnotherRelatedJenssegersItem extends JenssegersItem
+class AnotherRelatedItem extends Item
 {
     /**
      * @var string