+++ tests/_mocks/Items/Jenssegers/PlainJenssegersItem.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Swis\JsonApi\Client\Tests\Mocks\Items\Jenssegers;
-
-use Swis\JsonApi\Client\Items\JenssegersItem;
-
-class PlainJenssegersItem extends JenssegersItem
-{
-}