+++ tests/_mocks/Items/WithoutRelationshipsItem.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Swis\JsonApi\Client\Tests\Mocks\Items;
+
+use Swis\JsonApi\Client\Item;
+
+class WithoutRelationshipsItem extends Item
+{
+    /**
+     * @var string
+     */
+    protected $type = 'item-without-relationships';
+}